This is my first commit.
This is a battleship simulator where you can play against the computer or the computer can play against itself. You can choose the length of your board, difficulty, and whether or not you want to place your ships randomly. There are cheats which are fun to use. I can guarantee that you cannot break the code aside from intentional exits or errors to kick you out of the game. Make sure that character width is the same for all of them. Enjoy!